requirements.txt
pillow==10.3.0
pandas==2.2.2